requirements.txt
# requirements.txt
gitpython>=3.1.0
pyyaml>=6.0
networkx>=2.6.0
requests>=2.25.0
tree-sitter>=0.20.0